Refactor File Structure and Hard-Code API Key for Testing
The API Key is hard-coded to enable forked repo-related pull requests
to run all tests without a dependency on a secret environment variable.
We will plan to clean-up this directory every day and it will also be
constrained by the same size limitations as other non-paying clients
so the risk is low.